requirements.txt
pyTelegramBotAPI==4.14.0
google-generativeai==0.11.0
SpeechRecognition==3.8.1
pydub==0.25.1
requests==2.31.0